+++ banned.php
@@ -1,44 +1,34 @@
 <?php
 
-/**
- * banned.php
- *
- * @version 1.0
- * @copyright 2008 by ??????? for XNova
- */
-
 define('INSIDE'  , true);
 define('INSTALL' , false);
 
 $ugamela_root_path = './';
-include($ugamela_root_path . 'common.php');
-
+require 'common.php';
 
 includeLang('banned');
 
 $parse = $lang;
 $parse['dpath'] = $dpath;
-$parse['mf'] = $mf;
 
+$bans = doquery('SELECT * FROM {{table}} ORDER BY `id`', 'banned');
+$bans_count = mysql_num_rows($bans);
 
-$query = doquery("SELECT * FROM {{table}} ORDER BY `id`;",'banned');
-$i=0;
-while($u = mysql_fetch_array($query)){
-	$parse['banned'] .=
-        "<tr><td class=b><center><b>".$u[1]."</center></td></b>".
-	"<td class=b><center><b>".$u[2]."</center></b></td>".
-	"<td class=b><center><b>".gmdate("d/m/Y G:i:s",$u[4])."</center></b></td>".
-	"<td class=b><center><b>".gmdate("d/m/Y G:i:s",$u[5])."</center></b></td>".
-	"<td class=b><center><b>".$u[6]."</center></b></td></tr>";
-	$i++;
+$row_template = gettemplate('banned_row');
+while ($ban = mysql_fetch_row($bans)) {
+    $ban['name'] = $ban[1];
+    $ban['reason'] = $ban[2];
+    $ban['from'] = gmdate('d/m/Y G:i:s', $ban[4]);
+    $ban['to'] = gmdate('d/m/Y G:i:s', $ban[5]);
+    $ban['by'] = $ban[6];
+
+	$parse['banned_rows'] .= parsetemplate($row_template, $ban);
 }
 
-if ($i=="0")
- $parse['banned'] .= "<tr><th class=b colspan=6>Il n'y a pas de joueurs bannis</th></tr>";
-else
-  $parse['banned'] .= "<tr><th class=b colspan=6>Il y a {$i} joueurs bannis</th></tr>";
+if ($bans_count) {
+    $parse['bans_count'] = $lang['ban_thereare'] . ' ' . $bans_count . ' ' . $lang['ban_players'];
+} else {
+    $parse['bans_count'] = $lang['ban_no_bans'];
+}
 
 display(parsetemplate(gettemplate('banned_body'), $parse), 'Banned');
-
-// Created by e-Zobar (XNova Team). All rights reversed (C) 2008
-?>