+++ language/en/banned.php
@@ -2,10 +2,10 @@
 
 $lang['ban_title']    = "List of Banned Players";
 $lang['ban_name']     = "Username";
-$lang['ban_reason']   = "Ban Reason";
+$lang['ban_reason']   = "Ban reason";
 $lang['ban_from']     = "Banned from";
-$lang['ban_to']       = "Ban to";
+$lang['ban_to']       = "Banned until";
 $lang['ban_by']       = "By";
-$lang['ban_no']       = "Ban no";
+$lang['ban_no_bans']  = "There are no banned players";
 $lang['ban_thereare'] = "There are";
-$lang['ban_players']  = "Banned Players";
+$lang['ban_players']  = "banned players";